requirements.txt
# external requirements
click
coverage
python-dotenv>=0.5.1

cupy-cuda101==7.4.0
chainer==7.4.0
chainercv==0.13.1
sklearn
h5py
parse
imageio
seaborn
gensim
yacs==0.1.7
neptune-client==0.4.119
